Refactor stories into their own modules
