+++ .storybook/config.js
@@ -1,7 +1,12 @@
-import { configure } from '@storybook/react';
+import React from 'react';
+import { configure, addDecorator } from '@storybook/react';
+import Container from '../src/component/container';
+
+addDecorator(story => <Container>{story()}</Container>);
 
 function loadStories() {
-  require('../stories');
+  const context = require.context('../stories', true, /\.js?$/)
+  context.keys().forEach(context)
 }
 
 configure(loadStories, module);
