+++ stories/screen.js
@@ -1,0 +1,5 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+import WelcomeView from '../src/view/welcome';
+
+storiesOf('Screens', module).add('Welcome', () => <WelcomeView />);
